Fixed loads of shit. Programs actually run now.
Check the diffs for more info. Most of the changes are self-explanatory.
Program currently performs very, very badly in the benchmark program, and
of course it's because of the allocate linear scan. Had to increase the
number of available arrays to 2^32, and Sandmark allocates rapidly.

So, basically, need to create an array structure with bult in size so that
the whole array of indexes crap can get nixed.